sanity checks to don't overwrite existing files

+++ src/bin/kcache_extract.rs
@@ -5,6 +5,7 @@ use clap::{App, Arg};
 use std::error::Error;
 use std::fs::File;
 use std::io::prelude::*;
+use std::path::Path;
 
 fn main() {
     let matches = App::new("kcache_extract")
@@ -42,6 +43,11 @@ fn main() {
     let input_filename = matches.value_of("input").unwrap();
     let kernelcache_output_filename = matches.value_of("output").unwrap();
 
+    if Path::new(kernelcache_output_filename).exists() {
+        println!("file {:?} already exists", kernelcache_output_filename);
+        return;
+    }
+
     match kcacheext::extract_from_file(input_filename) {
         Ok(decoded) => {
             let mut kernelcache_output_file = match File::create(kernelcache_output_filename) {
@@ -58,6 +64,10 @@ fn main() {
             };
             if matches.is_present("kpp") {
                 let kpp_output_filename = matches.value_of("kpp").unwrap();
+                if Path::new(kpp_output_filename).exists() {
+                    println!("file {:?} already exists", kpp_output_filename);
+                    return;
+                }
                 if decoded.kpp_present {
                     let mut kpp_output_file = match File::create(kpp_output_filename) {
                         Err(why) => panic!("couldn't create {}: {}", kpp_output_filename, why.description()),
